Added BQML training method
